Update #1
Backend done and have figured out how to interact metamask/ganache  and contract with React. Have done a bit of html on displaying account address and detecting account changes. In the next week, I can probably be done with most of the html and start working on css. Definitely could be done by the first/second week of Winter quarter.